requirement.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.8.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
curl_cffi==0.11.1
Django==5.1.7
dnspython==2.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.0
fastapi-cache2==0.2.2
fasteners==0.14.1
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozendict==2.4.6
future==0.18.2
greenlet==3.2.3
h11==0.16.0
httplib2==0.20.2
icoextract==0.1.2
idna==3.10
importlib-metadata==4.6.4
itsdangerous==2.2.0
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==3.0.2
monotonic==1.6
more-itertools==8.10.0
multitasking==0.0.11
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
paramiko==2.9.3
passlib==1.7.4
peewee==3.18.1
pefile==2021.9.3
pendulum==3.1.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.8
protobuf==6.31.0
psutil==5.9.0
psycopg==3.2.6
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pycairo==1.20.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyelftools==0.27
Pygments==2.11.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==5.4.1
redis==6.2.0
reportlab==3.6.8
requests==2.32.3
rsa==4.9.1
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlalchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
texttable==1.6.4
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yfinance==0.2.61
zipp==1.0.0
